Moved all refs to template level to be under on scroll listener
