+++ src/components/sections/About/Background/index.js
@@ -51,7 +51,6 @@ export const backgroundHandler = refs => {
   titleOne.style.transform = `translateX(calc(${scrolled}px - ${offset}%))`;
   titleTwo.style.transform = `translateX(-${scrolled}px)`;
   titleThree.style.transform = `translateX(calc(${scrolled}px - ${offset}%))`;
-
   if (width < 768) {
     compRef.style.transform = `translateY(-${scrolled}px) rotateZ(-5deg)`;
   } else if (width < 959) {
