+++ src/global/utils.js
@@ -121,3 +121,10 @@ export const maintainRatio = (width, height) => `
     }
   }
 `;
+
+export const setPayload = (refs, setRefs) => {
+  refs.forEach(ref => {
+    const payload = { payload: ref, type: 'create' };
+    setRefs(payload);
+  });
+};
